Changes to LinesPanel, would like @smelc to check
LinesPanel didn't work well when I first tried to swap out a
SquidMessageBox for a LinesPanel, and I changed several things so the
behavior is more uniform. First, addLast() would add at the end but
(maybe due to copy-paste errors) would remove from the end as well,
counter to what you'd expect for a scroll, that is, it should remove
from the other end from where you insert. This behavior is kept by the
new setLast() method, in case it's needed. Second, draw() needed a lot
of changes, some of them to account for fonts that can be resized (now
LinesPanel can take a distance field font via a TextCellFactory, though
it won't create actual cells). A key change here was to make the line
height more exact; before, if you added a line that consisted of the
String "...", it would have lower height than a string that contained
"ALL CAPS", and "mama" would have lower height than "papa" due to the
GlyphLayout's height being used in some cases when it can fluctuate a
lot, including to greater than lineHeight. This does change the way
height is calculated, and it is now exactly the number of lines times
the lineHeight, but doesn't account for the bounds of the LinesPanel,
which had previously been a problem when the window could resize and
stretch existing Actors. If drawBottomUp is false, the default, then
some things will probably be in different places in this version, since
they seemed to be drawn off-screen with the previous way. I'm not sure
why this was, exactly.