Add Horde and Herd RNGs, like longer-period LapRNG
Horde is like LongPeriodRNG but probably not quite as long-period and a
fair bit faster thanks to a similar design to LapRNG (Horde is close in
speed to ThunderRNG, sometimes). Herd is like Horde but uses int math
only, so it should be expected to have a lower period but be faster on
GWT and 32-bit Java installations. These two RandomnessSources have
unusual performance characteristics, with their next() method seeming
unusually slow and usage from RNG slower as a result. However, usage on
their own is very fast (though not as fast as LapRNG or FlapRNG) and
their period is very good (it isn't clear just how good, but it should
be quite a lot better than Lap or Flap; 128x longer period than the
similar kind of generator would be on the low end of the estimate).