Add ZapRNG (LapRNG with quality up, speed down)
LapRNG already has perfectly decent quality for some uses, but ZapRNG is
better on quality while still being faster than LightRNG. It should be
comparable to ThunderRNG on speed and probably period, but with better
quality (and importantly, it shouldn't have ThunderRNG's issue of low
period for less significant bits). Unfortunately, this particular
generator type is incompatible with the nice skip() method in
PermutedRNG, PintRNG, and LightRNG, due to multiple sequences being
possible depending on the relationship of starting seeds.