Changes to Noise.LayeredXX, world map stuff
Noise.Layered2D and the other dimensionalities previously did not match
the normal expected behavior for octaves of noise, and would always have
bias toward a faint misty appearance. This had been worked around before
by using at most 2 octaves in HashVisualizer, but now is addressed
directly by fixing the math quirks in that and some other nested classes
of Noise.

DetailedWorldMapDemo currently is set up to output images that can be
reasonably used as textures for 3D models, likely to be used for
texturing spheres or similar shapes. Rivers and lakes are not shown on
these images. These will be produced in the out/worlds/ subdirectory of
the project root if you run the demo as it stands in this commit, and
out/ is ignored by git now to avoid polluting any commits. Each world
image is named with a random name from FakeLanguageGen. One sample world
image, Ark.png (a random name), is in the docs folder to be seen on
GitHub Pages as an example of what the spherical world map technique
looks like now.