BirdRNG is a high-quality, high-period int RNG
BirdRNG is very high-quality considering its speed, and passes all of
the PractRand suite for 64MB of random int data except for one optional
"folding mode," where it has a relatively low amount of failures, and
they have low severity. BirdRNG has a relatively large state but is
still fairly fast; it is much faster than LongPeriodRNG at generating
ints (which makes sense, since it's using int math for GWT
compatibility), but also competitive on long generation against that
generator. Period should be high, but just how high is unclear. It could
be below 2 to the 100, but this seems unlikely; it is almost certainly
higher than 2 to the 64, and cannot be higher than 2 to the 1056 (2 to
the 1024 is a more realistic upper bound on period). The period should
be high enough, certainly.

CrossHash.Mist also was adjusted here so it can't be mutated in-place to
change future hashes; this wasn't used in SquidLib but could have been a
problem if misused in other code. Mist.randomize() now returns a new
Mist instead of changing itself. There are also 48 prepared Mist
instances now instead of 24, since BirdRNG likes to have a at least 32
of them in a predefined array.