Add length limits to Bresenham, fix bug in DDALine
Bresenham now takes optional limits on the size of the returned Queue or
array, producing a smaller amount of Coords if it reaches the target
first, but guaranteeing no more Coords will be used than maxLength (It
stops short of the end target if maxLength can't reach).  DDALine had a
bug introduced at some point where it would only produce orthogonal
straight lines, and would rarely get to the target; this is fixed. This
is all tested in LOSComparisonTest, which was updated to use more modern
SquidLib code.