Add alterCarefully to OrderedMap
This method, along with the similar alterAtCarefully, will not change
the OrderedMap if the replacement key is already present. Its earlier
counterpart, alter(), can remove keys and values from the Map if the
replacement would duplicate an existing key (it replaces both original
and its duplicate with one entry). This is tested in OrderedDataTest.
This behavior isn't usually a problem unless code wasn't checking for
the presence of the replacement it was using, but because the docs
weren't very good, that wasn't fair to assume people would do.