Access to names made in PoliticalMapper; test!
PoliticalMapper wasn't tested (in SquidLib itself) before, and now
there's an easy test to check that it works. PoliticalMapper now also
stores a briefAtlas of just the culture-name parts of each nation, such
as "America" as opposed to the whole "United States of America". This
needed some changes to Thesaurus in how it generates nation names, but
now it also won't generate 1-or-2-letter names, which are much more
likely to be identical in large groups of nations.