Fix bug (?) in our impl of Isaac
At least in the 64-bit version, IsaacRNG, some of the numbers in regen()
should have been changed to match the official 64-bit ISAAC C version,
but were still the same as in the 32-bit ISAAC C version I was porting
initially (Isaac32RNG seems fine). Performance seems to be a tiny bit
better than before due to micro-optimization, but it's still a fairly
slow generator.