Change neighbor behavior, add GameOfLife
GreasedRegion's neighbor-finding functions behaved strangely when I was
implementing Conway's Game of Life functionality for CellularAutomaton,
and this was because they required two adjacent cells to be "on" instead
of just the neighbor. Most uses of neighbor-finding don't often require
an adjacent pair, for instance getting the shadow cast in one direction
from an area (which frequently would only contain cells that were "off"
in the original). This was changed for consistency and also to fix the
Game of Life code, which can't have neighbors required to be pairs.
HashVisualizer has a mode to visualize the Game of Life with black and
white pixels.