Add formCurvedFloat to NumberTools
It uses an existing long (or two ints), which probably would be random,
and gets a float between -1f and 1f, strongly biased toward 0f. A
precise use hasn't come up for it yet in SquidLib, but I found that
someone else's problem (related to LibGDX) might be solved more easily
with a method like it.