Improve NumberTools.formCurvedFloat
Now it gets closer to a Gaussian distribution in shape (actually closest
to Irwin-Hall), without needing more numbers (it can take just one int
now, if needed).

The asset files for the Light weights of Iosevka have been optimized for
file size, and the assets zip has been updated with the new files.