Add BardRNG, a high-quality GWT-optimized RNG
BardRNG passes PractRand with no anomalies given 64 MB of random values,
which is very good and not many generators do that here (several have no
failures, but having no anomalies is somewhat uncommon). BardRNG is
optimized for GWT by having a super-sourced version that uses typed
arrays; it's still slower than PintRNG or FlapRNG, but has a much much
higher period than PintRNG and vastly better quality and period than
FlapRNG (which is very flawed, but fast). Period should be higher than
BirdRNG and on par with BeardRNG; this has the same size of state as
BeardRNG. On GWT, speed is on-par with HerdRNG at about 4 times slower
than FlapRNG, about half the speed of PintRNG, and faster than any
long-based generators. LightRNG is about 9 times slower than BardRNG,
but BeardRNG is surprisingly only about 3-4 times slower, and it seems
Beard is the fastest of the long-based generators on GWT for no apparent
reason. Without super-sourcing, BirdRNG and BardRNG are extremely close
on performance, but with it, Bard takes 220 ms to generate 10 million
ints while Bird takes about 320 ms. The super-sourcing to use typed
arrays seems worth it, and it isn't too much effort.