POSSIBLY BREAKING: SquidLayers palette removed
This was an oddity in the API for some time, and the palette was only
needed so squidlib-util could pass some kind of code for a different
kind of Color to AWT or to libGDX without instantiating one. This
doesn't make sense though, since the indices only make sense to
something that can use them, so the color assigning code is being moved
to MapUtility in the squidlib display module, where it produces 2D Color
arrays instead of 2D int arrays. A whole lot of old demo/example code
needed to change to fit this, but I suspect the rigidity of the palette
approach as it was used in SquidLayers made it a poor fit for real
games. If you have a palette (the original was just
SColor.LIMITED_PALETTE wrapped in a List), then you can get the
appropriate Color from it and pass it to any method that takes a Color,
instead of using an int index to some default palette with that Color.