Weird fix to ThinWallDemo, make recallActor robust
Previously, ThinWallDemo was not rendering correctly, but this was a
simple enough fix once I figured out which modified dungeon map to use.
Some changes to recallActor have it producing automatically-handled
Actors in some cases, such as for summon() and burst(); earlier, if the
actors went off-screen, they would stay active and their animations
running but could not be fully recalled, possibly leading to lots of
invisible Actors using up processing time (for their Actions or other
animations) and RAM. This has been fixed, it looks like, with no change
to existing code that already worked (such as ThinWallDemo).