Various changes to WhirlingNoise, world gen
This involved two added static methods to LightRNG for quick calculation
of random 32-bit values from an int for state, and for random bounded
numbers for small bounds (they should fit in a short) given an int
state. These do well on GWT, unlike LightRNG normally. These are now
used in WhirlingNoise since we only had ints to start with anyway, but
there are other changes to WhirlingNoise. The 3D gradient vectors are
now based on an irregular-seeming polyhedron that has few re-occurring
straight lines when expanded to cover a sphere (the polyhedron is a
pentagonal hexecontahedron); this tendency of having few patterns is
desirable in noise. The 4D gradients were more of a source of problems,
with lots of long straight lines in the tiling mode of
WorldMapGenerator, and it seems to be set up now to avoid that much of
the time.

Iosevka's license is slightly updated for the case where someone removes
some files, so the license wouldn't cover them.