Some BlazingChain code -> SquidLib, add "crypto"
This removes the dependency on BlazingChain, which may make it easier to
rely on squidlib-extra for gdx-setup. What's been added is pretty shoddy
crypto, but it should work better than Garbler at making things like
savefiles impenetrable to the average player. Only two functions from
BlazingChain are moved over, and they have been simplfied somewhat to
only handle UTF-16 encoding (as is used in Preferences) instead of
needing to decide which code for any of four strategies to use. They are
in LZSPlus in squidlib-extra, and they have the previously-mentioned
crypto-lite code worked into the compress and decompress methods.
Garbler is still used when compression is off but garbling is on, for
some reason. The garble keys we use are int arrays now, which makes more
sense for varying strengths desired.