Change shader for distance field; usable small!
The size of anything rendered with a distance field font won't change as
a result of this, but if you scale down a window that has distance field
fonts in it, now it won't have the unpleasant glitchy appearance that
the fonts previously had at very small sizes. It still has a minimum
readable size, but it's smaller. Sample: http://i.imgur.com/gUmtzOx.png
. Thanks to @mfiano for finding the nice technique at
http://www.java-gaming.org/index.php?topic=33612.msg316185#msg316185