Add Light32RNG, add cross-module links in docs
Light32RNG is like LightRNG but int-optimized, for GWT or 32-bit
platforms. It has a much larger period than a pure SplitMix32 generator
(LightRNG is a pure SplitMix64 generator that has a period of 2 to the
64, using long math, but this uses some tricks to get a period of 2 to
the 63). Despite a larger period, it is very close to LightRNG on speed
for generating ints, and sometimes outperforms it on int generation.
Quality is extremely similar to LightRNG, getting no failures on testing
and the same number of "unusual" anomalies.

The Maven plugin dependencies have been updated, which only matters if
you use Maven and should automatically download the current versions.
This includes updated compiler, source, and javadoc plugins, and the new
JavaDoc configuration includes links across modules.