Utilities in data structures. Arrangement stuff.
The utilities are mostly Java 8 API additions to the Map interface,
which we can implement on Java 7 as long as we don't use the Override
annotation. The "Arrangement stuff" is the removal of addAllIfAbsent
methods, because they act the same as the existing putAll overloads for
Collections of K and for K arrays. They are not removed from
MultiArrangement because they have a reason to exist there.