Add padding methods in StringKit, invert() in SCC
... and previews for both DawnBringer palettes and the Color Wheel
palette in SColor, using code that produces the previews usng the new
methods. There's a slight refinement of the API for IColorCenter;
getHSV() should be able to handle corner cases a bit better since hue is
fundamentally a wrapping value, and inputs for hue that are lower than 0
or higher than 1 make sense.