Reduce allocations for lightness arrays with...
SquidLayers.autoLight(), which is similar to the
generateLightnessModifiers method in DungeonUtility or MapUtility but
doesn't produce a new 2D int array, instead reusing the existing
(previously wasted) lightnesses field in SquidLayers. SquidLayers now
actually reads this field instead of just writing to it, and unless it
is all 0 (which it will be if not set), it will affect the lightnessses
of backgrounds similarly to how it could be done manually before (and
still can be).