Fix distance field shader for Android, other opts
The distance field shader used dFdx and dFdy, two methods that are part
of GLSL on desktop and GL ES 3.0.0 and newer, but need an extension
(that is common) on GL ES 2.0 hardware. The new shader requires this
extension, which should be present on even barely-gaming-capable
devices.

The other opts are discussed in detail here,
https://shipilev.net/blog/2016/arrays-wisdom-ancients/ ; I took any
`Collection<T>.toArray(T[])` calls that allocated an array with empty
contents to be filled by toArray, and now pass a zero-length array to
toArray so it can resize it without needing to "zero-fill" the older
large arrays before filling them again.