Fix distance field font shader for GWT, also Android
The last commit worked on most Android devices but not GWT. This commit
avoids using dFdx, dFdy, and fwidth entirely, using some somewhat-rough
approximations that seem to work in all the fonts in FontTest, while
also (in theory) being a little less work for the GPU to do. The MSDF
shader changed as well, and may look slightly different.