Changes to CrossHash to make hashing cross-platform
This mainly has an impact if you run an application with the same seeds
on desktop or Android and GWT; before, using hash64() would generally
work the same across platforms, but Wisp, Mist, and Falcon would behave
differently on GWT due to using 32-bit math, which was assumed to be OK
because JS doesn't have native 64-bit integers. I was wrong; because JS
imitates 32-bit ints only for bitwise operations, not multiplication or
addition, many values were exceeding 32 bits and producing strange
results in code that uses 32-bit math, including CrossHash.Wisp,hash(),
BardRNG, and Light32RNG, the last two of which were designed for use on
GWT. If you use a 64-bit-math RNG (such as the default LightRNG, the
faster-and-lower-quality LapRNG, or the higher-period XoRoRNG), then you
should have the same results on desktop and GWT, but not if the RNG
relies on 32-bit int overflow or precision with large int values (which
have trouble beyond 2 to the 53 due to being doubles on GWT).  Using
CrossHash.hash64() now uses Wisp instead of the old-bad FNV-1a
implementation, and CrossHash.hash() simply truncates hash64()'s result
to get an int, which is slower on GWT but reliable in its quality and
behavior between platforms. The tests will need some naming updated for
this change.