The long path to move more code to 64-bit math
Because GWT treats longs reliably and correctly, but doesn't necessarily
do so with ints, I'm moving as much cross-platform code as possible to
use longs when random numbers, hashing, or other usages that may rely on
integer overflow is involved and it should act the same on all
platforms. This commit starts that with Garbler and LZSPlus, used by
squidlib-extra.

TinyTest2 has some changes to try to diagnose what the problem could be.