Start work on a better way to render colorful chars
SquidPanel does well when it has a dense, solid map where all characters
are visible and have a color, but when multiple layers are involved,
many of them won't have chars in them and will still need `width *
height` chars rendered as invisible space. Using a sparse storage for
chars and their colors makes sense for many kinds of map, and we're
starting to get that with SparseTextMap.

This commit also changes the hashCode implementation for a few classes
that would have had issues on GWT with consistency, including Coord and
the internals of StringConvert, where it probably matters much less than
for Coord. For code using ordered collections, the change in hashing
shouldn't have any noticeable changes.