Much faster simple rendering with SparseLayers
It should scale in the future to handle more complex cases, like sliding
Actors over the map, as well. There's essentially no visible time
between frames when you move in SparseDemo, which is even better than
BasicDemo2, which it is based on.

Also, this commit goes back to using libGDX 1.9.6 stable.