Add wiggle() to SparseLayers; cleanup, fixing
SparseDemo had a small bug before where the pathfinding for the mouse
path wouldn't work correctly if you had moved with the arrow keys (or
WASD/HJKL). Since the new feature added to SparseDemo is that you wiggle
if you use the arrow keys to try to move into a wall, this bug was
noticeable, and the fix turned out to be pretty easy. There's also some
cleanup to avoid repetition in SparseLayers for how it translates from
grid to world coordinates.