Hackish solution to #180 seems to work
There's a small amount of replicated code in summon() now, but it needs
to do some things only after the delay has elapsed, including creating
the Actor that the rest of the effect acts on (otherwise the colors
don't work correctly). This also seems to have no problems with burst().
You can test in tests/BasicDemo2 by using the keyboard (arrow keys, hjkl
or wasd) to bump into walls, which causes a fading-out '?' to appear. I
also tried with fading-in and the bug seems gone from both. This doesn't
investigate the slide issue mentioned in #180 , and that may come later.