SparseLayers.bump(); help SquidPanel.bump()
SparseLayers.bump() does what it is expected to do, and makes a bumping
effect for a Glyph to push into and back from a specified Direction.
SquidPanel.bump() may have had a lingering bug for a while because it
didn't try to bump a fraction of the full width or height of a cell, but
a fraction of a world coordinate (here, usually a pixel). It also was
set up to use two different alignments for the Actor, for some reason.
The SquidPanel one should be improved, if not fully fixed. This also
adds Direction.getRoughDirection(), which can quickly get the
appropriate direction for x and y differences when they are small, such
as the common case of moving one cell orthogonally or diagonally and
wanting to get the Direction for that move.