Various tweaks to get explosions round
This involved a change that affects radius.pointsInside() when used with
a CIRCLE or SPHERE value only, and should allow that method to produce a
more round circle more of the time. PanelEffect.ExplosionEffect wasn't
calculating the most distant cells of the explosion accurately, so parts
were just getting truncated; now the FOV call it uses just has a larger
radius by a half-cell.