Tricky fixes to conical FOV
This affects the previous commit, which didn't work on some angles, and
also the earlier FOV method that used an angle and span. Both would
sometimes consider no cells as seen when the angle was too close to
certain values, like 0, for odd reasons probably related to math
mistakes. They've both been tested on quite a lot of angles now, and
seem to be correct. It's possible there are still bugs with span, but
they would need to be brand-new, since span was fine before.