Fix issue with WhirlingNoise in 3D/4D, demo Cosmic
WhirlingNoise wasn't always producing numbers in-range for -1.0 to 1.0
in 3D and possibly 4D, and this was probably the result of the shape it
uses for its pseudo-random vectors being slightly larger than I
expected; it's been shrunk down and the artifacts from too-high-or-low
numbers are gone. CosmicNumbering is demonstrated as a noise function in
HashVisualizer, though it isn't really a noise function in the
traditional sense and doesn't act too well as one.