Add wider Iosevka to assets; adjust distance field
The new distance field fonts added are all Iosevka with 15% wider
characters at the same consistent stroke weight, but come in slab-serif
and sans-serif variants as well as light and regular-weight variants.
They are all available in DefaultResources with "StretchableWide" in the
name of the method. All distance field fonts (except MSDF) now have the
shader take into consideration the vertical and horizontal stretch of
the font when determining how much to sharpen or smooth the font's
appearance; before they only considered horizontal stretch, which could
cause the font to be blurry if it was mainly stretched vertically. The
assets zip file has been updated.