Improve BasicDemo2, fix some bugs
BasicDemo2 now uses a smaller block of code that's more
technically-impressive (and educational) for the language stuff it does;
now it shows 6 lines from Sun Tzu's The Art of War, then ciphers them
using NaturalLanguageCipher into various randomly-selected fake
languages.

Bugs fixed: OrderedSet.toArray(), when passed a zero-length array, would
not and (probably?) cannot resize the array on GWT, where GWT would use
an internal section of code to accomplish the resize on its emulated JDK
collections. So even though a zero-length array is faster on desktop and
for JDK collections like ArrayList, we need to use a correctly-sized
array with OrderedSet. There was a bug in StringKit.wrap() when passed
non-ASCII words, such as text in the Greek alphabet; this stems from a
bug in RegExodus but can be easily worked around in this case (I needed
to change `(?=\\S)` to `(?!\\s)`, how much clearer could it get...
grumble...).

NaturalLanguageCipher now has an easier way to reset the whole cipher if
the language or shift (used for dialects) needs to change. A small
possible bug was fixed in its internal StatefulRandomness used to
less-randomly generate words from similar initial words, with similarity
gauged by a phonetic hash.