Improve SparseDemo by showing viewports and langs
The big block of FakeLanguageGen.SentenceForm objcts is gone, replaced
with a less verbose example using NaturalLanguageCipher and randomly
selected languages from FakeLanguageGen (earlier added to BasicDemo2).
The translations start with a passage from Sun Tzu's The Art of War, and
translate it to various fake languages, showing the current text in a
separate viewport below the main map.

This also adds a method to SparseLayers that can clear a single
foreground layer, and checks if the TextCellFactory the constructor is
given has already been initialized (useful when two SparseLayers share a
TextCellFactory and one has already initialized it).