Bunch of convenient methods in SparseLayers
A bug was fixed in SparseLayers (width and height confusion in an
infrequently-used section of put(char[][], Color[][]) ). Documentation
is also better in SparseLayers, though it was good before. Also an
addition to SquidColorCenter.rainbow() so it can set the opacity on the
colors it makes (this isn't used currently, but was earlier in the
development of this commit). SparseDemo has some other small
improvements.