Improve SColor docs, close #182
More than just the DB and CW colors were lacking consistent `@`
colorization; there were cases like that in most parts of SColor. A new
block of code in ColorTest generates docstrings that contain the name,
RGB color code, hue, saturation, and value of each color, on top of
fixing the issue with colorization and an unreported bug that may only
affect IDEA, where the second row didn't appear because <p> isn't the
same as <br>. The docstring writer is potentially extensible to produce
more info if we need it, like giving red/green/blue as floats.