Opts and fixes to Zone, GreasedRegion
GreasedRegion.toArray() with no arguments would return an
incorrectly-empty Object array every time; now it delegates to
asCoords(). On the other hand, toArray with a T array argument wasn't
assigning to the contents of the argument, and that's fixed now. There
is now an overload of asCoords() that takes an existing Coord array; it
will fill as much of it as it can and stop short if the array is too
small. The existing asCoords() method now delegates to the new one,
giving it a correctly-sized array for the full contents. There's some
bounds checking in nth(). More (most?) Zone methods are implemented in
GreasedRegion, with some simply being able to call existing methods, and
some benefiting from more involved code (such as x(), which can avoid
iterating over all points by taking advantage of the internal structure
of the GreasedRegion). Zone.Skeleton now marks width and height as
protected, not private, so they can potentially be used by subclasses
(GreasedRegion may have a use for them, and can access them by
super.width or super.height). Lots of docs added to Zone, specifying the
8-way adjacency that some methods use (this is also in DungeonUtility
for its border() method).

Splash had an odd spelling mistake in two method names, one of which was
protected and the other private. If you extend Splash and override
"drunkinize", you should change it to "inebriate", which also means
"become drunk." I'd prefer using "inebriate" because it doesn't make it
look like the name refers to something the programmer is doing when they
implement "drunkinize".