Simple method of cave generation with CA
This isn't the standard means of cellular automata cave gen, where the
original state is made simply of randomly scattered walls, but instead
it uses a room-and-corridor map from DungeonBoneGen and the other
TilesetType code. This produces some rather complex layouts thanks to
the TilesetType behavior, but should look very cave-like or organic.