Add FakeLanguageGen.ALIEN_E
This alien language works for alien species that lack lips (they may
have beaks), since it avoids 'p', 'b', and 'm' sounds, but uses many
kinds of click noises to make up for the missing phonological
repertoire.  The idea here is to eventually have an ALIEN_ variant for
A, E, I, O, and U, where the letter is the dominant vowel used more than
any other.