Adjust WhirlingNoise to use better gradients
This has only been tested for 3D so far, in DetailedWorldMap(Writer and
Demo), but 4D is affected as well. This seems to have somewhat better
behavior regarding artificial-seeming straight lines in world maps,
though it wasn't especially bad before, and has a cleaner (maybe also
faster) implementation because the shape used has 32 vertices (a power
of 2) instead of 92. The shapes tend to look more curvy and less regular
here, so that's the main goal I was trying to obtain.