Add variant on demo's biome code to -util
The biome-gradient code should be useful in more places than just a
demo. It may be a good next step to predefine the same colors the demo
uses as named colors in SColor (in part because they're good natural
colors, but mostly for use with world maps), and organize them in a
palette that this can be used with.