Fix broken behavior in StringKit int/long reading
I'm not sure when this regressed, since it was fine earlier this year,
but the numbers returned were inaccurate when I used StringKit in
another project. I created a unit test, which confirmed the problem and
that it affected the intFromHex, intFromDec, longFromHex, and
longFromDec methods. The fix turned out to be fairly simple; the first
char was being read and used but then its position was being read and
used again. StringKit also now has a general-use join() method that
takes an Iterable of any object type (though it won't work well with an
Iterable over array items) and calls toString() on each object iterated
over to get the text to join (separated by a delimiter). This is also
tested.