Key fixes to DetailedWorldMapWriter, biome stuff
The biome code used for oceans was incorrect and caused array bounds
exceptions when the biome was requested for an ocean. The code for
fetching biomes is better now, since the data that is given for a biome
code is not just a number 0-13, and actually can range from 0 to 59 to
allow for alternate biome tables. The generated code also includes a
simple way to fetch the name of a biome type for a cell using the
default biome table.