RNG SEED RESULT CHANGE: Shuffle improvements
Now, the various shuffle(), shuffleInPlace(), and randomOrdering()
methods all use the same algorithm (with slight adjustments to fit some
cases), and now shuffling an array like {0, 1, 2, 3} with a seeded RNG
will give the same result as if you had called randomOrdering() with the
same RNG. The shuffling methods used a hodge-podge of various algorithms
before, some of which may have been less correct. They also made an
unpredictable amount of calls to the next() or nextLong() methods in
that RNG, which has implications if a shuffle needs to be reversible
using something like LightRNG.skip(). Now the number of nextLong() calls
should be stable at (arrayLength - 1), where arrayLength is the number
of items being shuffled or the length parameter passed.