Help SparseLayers.recallActor() during an Action
I'm not sure if this is a complete fix.  There's issues in another
project that uses SparseLayers and could overshoot the bounds of the
glyphs list when a Glyph is recalled during its action. This seemed to
be introduced when I switched away from using a SnapshotArray to just
use an ArrayList, since some code wasn't updated.