Update CosmicNumbering with a less-glitchy look
CosmicNumbering uses a simpler technique than before that yields very
different-looking results (see HashVisualizer; TabbyNoise is the same
type as the new technique in 3D, and looks like
https://i.imgur.com/6pibn71.gifv , while CosmicNumbering also has a test
much later in the noise section).This also fixes a compilation error
introduced in the last PR, where GreasedRegion couldn't access the width
and height variables in Zone.Skeleton. These are back to being
protected, since I don't see much point in having private variables in
an abstract class, especially if those variables would often be
duplicated in extending classes, and even still can be duplicated if
they need to be.