Update TabbyNoise to use the seed more, reduce bias
In DetailedWorldMapDemo, you can press 't' to switch to using
TabbyNoise, 's' to switch to SeededNoise (the default from a while ago),
or 'w' to switch to WhirlingNoise (the current default, and probably the
best for map gen that we currently have). DetailedWorldMapWriter
currently writes small seamless maps using TabbyNoise, where it does
fairly well.