Significant update to CW_ colors, add brown
Several of the CW colors were stuck at a slightly-wrong hue with the
earlier way of using selected hues at fixed points on the color wheel.
Now, each color can use a slightly different hue if it matches the
aesthetic, so CW_YELLOW isn't slightly-too-orange any more, and various
other colors are closer to desirable hues as well. CW_BROWN is a new
group of 9 colors at a hue between CW_ORANGE and CW_APRICOT, and changes
some of the indices in COLOR_WHEEL_PALETTE, but should be a good way of
adding in some comonly-used less-saturated colors to the palette.