New QRNG as an alternative to VanDerCorputQRNG
SaltyQRNG has some nice properties and I'm shocked I was able to find a
kind of QRNG that works and I hadn't already known about, without
consulting any papers on the topic. This also does some reorganizing in
CoordPacker to make the static initialization a little cleaner and more
efficient.