Adjustments to SaltyQRNG to move low-period bits
The salt can be larger now, which is good. This still has low-period
bits, but they aren't nearly as noticeable since they've been moved to
around the sixth-lowest bit instead of the lowest bit, and only in the
integer-like number generation (nextFloat() and nextDouble() are almost
the same, with a change to account for the larger salts).