Minor fixes in WaveletNoise
It was capable of infrequently generating numbers outside the expected
range of [-1,1), probably because it added the whole seed to x, y, and z
coordinates individually, which made things start to get less precise.
Also, trying to use uniformly-distributed values for `noise` does not
work well at all; we're back to using an almost-Gaussian distribution
for the noise array.