Complete TabbyNoise changes started in last commit
In TabbyNoise, 2D, 4D, and 6D noise are now updated to use "wiggly
lines", in addition to 3D noise. Also, the option to use precalculated
seed and random angle data is present, though undocumented, and is used
both internally in calls to getNoise(), and in the demo HashVisualizer
to speed up display slightly (this seems to compensate for a hit this
takes to speed when it tries to reduce the predictable straightness of
lines).