FakeLanguageGen cleanup for Modifier
FakeLanguageGen.Modifier.replacementTable(String, String) is now called
FakeLanguageGen.Modifier.charReplacementTable(String, String) to better
reflect its behavior of replacing one char with another char and never
changing the length. This allows
FakeLanguageGen.Modifier.replacementTable(String...) to be used like the
version that takes an OrderedMap, but instead uses an array or vararg of
Strings; this should be more convenient and slightly more efficient.
There are new methods that insert a new possible vowel or consonant into
the groups that a language can generate as a Modifier; these can also be
specialized to only affect closing or opening parts of a word, and have
the probality of using the new insertion configurable.
FakeLanguageGen.Modifier.SIMPLIFY_ARABIC is now marked as deprecated,
since it has been a no-op for some time, isn't ever necessary except for
compatibility, and just increases the verbosity of code that uses it.
SIMPLIFY_ARABIC has been removed from where it was used in SquidLib's
tests.