SquidInput can save mappings; SquidPanel.bump fix
The saved mappings are used in BasicDemo2 to quickly add (or remove by
commenting out) support for wasd and vi-keys (with the option for 8-way
movement). SquidPanel.bump() previously used the wrong vertical
direction for its animation; this isn't a problem in SparseLayers, so
the contrast is probably what made me notice and fix this.