Add CriticalRNG for super-luck, update others
CriticalRNG should make some unusual cases easier to implement, where a
character can exceed the normal limits of what some ability can do by
sheer luck. It isn't a general RNG since it can and frequently will
produce values that are larger or smaller than an arrays's limits if
it's used to get a random element from an array as-is; instead it could
be used to get multiple items from that array if the value is very high,
or from a different array if it is high or low.

Various subclasses of RNG hadn't been looked at in a while; some methods
were marked GwtIncompatible but that no longer applied as of GWT 2.8.0,
while others could be made more general thanks to changes in APIs
elsewhere (String parameters were just passed to CrossHash, so they
could be CharSequence parameters and so take StringBuilders and such).