Back to snapshots. Dice changes + tests
Dice now avoids allocating an ArrayList of Integer every time a best-of
dice roll is made (it reuses an IntVLA), and also allows worst-of dice
roll notation. The old `3:5d6` notation is still available, but the new
`3>5d6` notation is preferred for "highest of" and `3<5d6` can be used
now for "lowest of". A small bug was fixed in the regular expression
uses to match dice calls; quite a few chars were allowed in the last
operator position that shouldn't have been. Several things have been
changed to better reflect usage of Dice; `rollGroup(String)` is now
`roll(String)` and has its javadocs cleaned up, `independentRolls(int,
int)` returns an IntVLA instead of a List of Integer, and most of the
code that doesn't make sense as public methods is now private.